refactor: DetailTextコンポーネントを作成
