+++ web/src/app/(app)/users/[userId]/page.tsx
@@ -9,30 +9,7 @@ import { deleteUser, getUser } from "@/lib/api/user";
 import { User } from "@/types/user";
 import { useApiSubmit } from "@/hooks/useApiSubmit";
 import { toast } from "react-toastify";
-
-type PropsType = {
-  label: string;
-  value: string;
-};
-
-const UserDetailText = (props: PropsType) => {
-  return (
-    <div
-      style={{
-        display: "flex",
-        alignItems: "center",
-        borderBottom: "dashed",
-        marginTop: 10,
-        paddingBottom: 10,
-      }}
-    >
-      <div style={{ display: "flex", fontSize: 20, fontWeight: 600 }}>
-        {props.label}
-      </div>
-      <div style={{ display: "flex", paddingLeft: 20 }}>{props.value}</div>
-    </div>
-  );
-};
+import DetailText from "@/components/DetailText/DetailText";
 
 const UserDetailPage = () => {
   const [user, setUser] = useState<User | null>(null);
@@ -83,9 +60,9 @@ const UserDetailPage = () => {
       >
         {user && (
           <>
-            <UserDetailText label="ユーザID" value={user.id} />
-            <UserDetailText label="ユーザ名" value={user.name} />
-            <UserDetailText label="メールアドレス" value={user.email} />
+            <DetailText label="ユーザID" value={user.id} />
+            <DetailText label="ユーザ名" value={user.name} />
+            <DetailText label="メールアドレス" value={user.email} />
           </>
         )}
       </div>
